Refactoring and todos after merge
